// lib/astro/horoscope/remedy/types.ts
export type RemedyCategory =
  | "mantra"
  | "puja"
  | "daan"
  | "gemstone"
  | "lifestyle"
  | "other";
// ...
export interface ResolvedRemedy {

  /**
   * Remedy category.
   */
  category: RemedyCategory;

  /**
   * Normalized remedy title.
   */
  title: string;

  /**
   * Primary practical remedy practice.
   *
   * Usually sourced from procedure,
   * with description as fallback.
   */
  practice: string;

  /**
   * Human-readable remedy guidance.
   *
   * Usually sourced from description,
   * with procedure as fallback.
   */
  guidance: string;

  /**
   * Original/general CMS description.
   *
   * IMPORTANT:
   * This is knowledge data passed through
   * from RemedyKnowledge.
   *
   * It is NOT generated by the resolver.
   */
  description?: string;

  /**
   * Resolver metadata explaining
   * the planetary relevance.
   */
  reason: string;


  ////////////////////////////////////////////////////////////
  // MANTRA
  ////////////////////////////////////////////////////////////

  mantra?: string;


  ////////////////////////////////////////////////////////////
  // BENEFITS
  ////////////////////////////////////////////////////////////

  benefits?: string[];


  ////////////////////////////////////////////////////////////
  // PRECAUTIONS
  ////////////////////////////////////////////////////////////

  precautions?: string[];


  ////////////////////////////////////////////////////////////
  // AVOID FOR
  ////////////////////////////////////////////////////////////

  avoidFor?: string[];


  ////////////////////////////////////////////////////////////
  // SUITABLE FOR
  ////////////////////////////////////////////////////////////

  suitableFor?: string[];


  ////////////////////////////////////////////////////////////
  // MATERIALS
  ////////////////////////////////////////////////////////////

  materials?: string[];


  ////////////////////////////////////////////////////////////
  // DURATION
  ////////////////////////////////////////////////////////////

  duration?: string;


  ////////////////////////////////////////////////////////////
  // GEMSTONE
  ////////////////////////////////////////////////////////////

  gemstone?: string;


  ////////////////////////////////////////////////////////////
  // METAL
  ////////////////////////////////////////////////////////////

  metal?: string;


  ////////////////////////////////////////////////////////////
  // DAY
  ////////////////////////////////////////////////////////////

  day?: string;


  ////////////////////////////////////////////////////////////
  // COLOR
  ////////////////////////////////////////////////////////////

  color?: string;


  ////////////////////////////////////////////////////////////
  // MEDIA
  ////////////////////////////////////////////////////////////

  media?: {

    image?: string;

    video?: string;

  };


  ////////////////////////////////////////////////////////////
  // SOURCE
  ////////////////////////////////////////////////////////////

  source?: {

    slug?: string;

    planet?: string;

    zodiacSign?: string;

  };

}